requirements.txt
backports.functools-lru-cache==1.4
backports.weakref==1.0.post1
bleach==1.5.0
cycler==0.10.0
funcsigs==1.0.2
html5lib==0.9999999
Markdown==2.6.9
matplotlib==2.1.0
mock==2.0.0
numpy==1.13.3
olefile==0.44
pbr==3.1.1
Pillow==4.3.0
protobuf==3.4.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
scipy==1.0.0
six==1.11.0
subprocess32==3.2.7
torch==0.2.0.post3
torchvision==0.1.9
Werkzeug==0.12.2
